Server-side support for Raspberry Pis (#68)

+++ ltsp/server/kernel/55-kernel.sh
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 # Copy vmlinuz and initrd.img from image to TFTP
+# @LTSP.CONF: RPI_IMAGE
 
 kernel_cmdline() {
     local args
@@ -38,6 +39,12 @@ Please export ALL_IMAGES=1 if you want to allow this"
         img_path=$(add_path_to_src "${img_src%%,*}")
         img_name=$(img_path_to_name "$img_path")
         re test "kernel_main:$img_name" != "kernel_main:"
+        if [ "$img_name" = "$RPI_IMAGE" ] &&
+                [ -f "$BASE_DIR/$img_name/boot/bootcode.bin" ]
+        then
+            re rpi_image "$img_name"
+            continue
+        fi
         tmp=$(re mktemp -d)
         exit_command "rw rmdir '$tmp'"
         # tmp has mode=0700; use a subdir to hide the mount from users
@@ -64,6 +71,28 @@ EOF
     if [ "$runipxe" = "1" ]; then
         echo "To update the iPXE menu, run: ltsp ipxe"
     fi
+}
+
+# Symlink all $BASE_DIR/$RPI_IMAGE/boot/* files directly under $TFTP_DIR
+rpi_image() {
+    local img_name var dst
+
+    img_name=$1
+    echo "Symlinking $BASE_DIR/$img_name/boot/* in $TFTP_DIR/*"
+    re test -f "$BASE_DIR/$img_name/boot/bootcode.bin"
+    for var in "$BASE_DIR/$img_name/boot/"*; do
+        re ln -rsf "$var" "$TFTP_DIR/${var##*/}"
+    done
+    # Remove old, dangling symlinks
+    for var in "$TFTP_DIR/"*; do
+        test -L "$var" || continue
+        dst=$(re readlink -f "$var")
+        if [ "$dst" = "$BASE_DIR/$img_name/boot/${var##*/}" ] &&
+                [ ! -e "$dst" ]; then
+            warn "Deleting dangling symlink: $var"
+            re rm "$var"
+        fi
+    done
 }
 
 # Search for the kernel and initrd inside $dir
